networ, util: Move Is{Primary,Additional}Country into the network package
The previous commit introduced a import cycle:

```
package github.com/etix/mirrorbits
	imports github.com/etix/mirrorbits/cli from main.go
	imports github.com/etix/mirrorbits/filesystem from commands.go
	imports github.com/etix/mirrorbits/config from hash.go
	imports github.com/etix/mirrorbits/utils from config.go
	imports github.com/etix/mirrorbits/network from utils.go
	imports github.com/etix/mirrorbits/config from geoip.go: import cycle not allowed
```

This is because we now import `utils` from within `config`.

Let's move the two network-related utils to `network/utils.go`, so that
we don't need to import `network` in `utils.go` anymore. That's enough
to break the import cycle.
